# perf: optimize DOM traversal