This is like a 1% perf improvement in Firefox and 1-10% in Chrome for the `many-custom-emoji` benchmark, so seems worthwhile.

```
┌─────────────┬──────────────────┐
│     Browser │ firefox-headless │
│             │ 130.0            │
├─────────────┼──────────────────┤
│ Sample size │ 1990             │
└─────────────┴──────────────────┘

┌─────────────┬─────────────┬─────────────┬─────────────────────┬─────────────────┬─────────────────┐
│ Benchmark   │ Version     │ Bytes       │            Avg time │  vs this-change │  vs tip-of-tree │
│             │             │             │                     │                 │     tip-of-tree │
├─────────────┼─────────────┼─────────────┼─────────────────────┼─────────────────┼─────────────────┤
│ this-change │ <none>      │ 1027.21 KiB │ 254.85ms - 257.75ms │                 │          faster │
│             │             │             │                     │        -        │         0% - 2% │
│             │             │             │                     │                 │ 0.32ms - 4.40ms │
├─────────────┼─────────────┼─────────────┼─────────────────────┼─────────────────┼─────────────────┤
│ tip-of-tree │ tip-of-tree │ 1026.74 KiB │ 257.22ms - 260.10ms │          slower │                 │
│             │             │             │                     │         0% - 2% │        -        │
│             │             │             │                     │ 0.32ms - 4.40ms │                 │
└─────────────┴─────────────┴─────────────┴─────────────────────┴─────────────────┴─────────────────┘


┌─────────────┬─────────────────┐
│     Browser │ chrome-headless │
│             │ 129.0.0.0       │
├─────────────┼─────────────────┤
│ Sample size │ 260             │
└─────────────┴─────────────────┘

┌─────────────┬─────────────┬─────────────┬─────────────────────┬──────────────────┬──────────────────┐
│ Benchmark   │ Version     │ Bytes       │            Avg time │   vs this-change │   vs tip-of-tree │
│             │             │             │                     │                  │      tip-of-tree │
├─────────────┼─────────────┼─────────────┼─────────────────────┼──────────────────┼──────────────────┤
│ this-change │ <none>      │ 1027.21 KiB │ 172.61ms - 183.40ms │                  │           faster │
│             │             │             │                     │         -        │         1% - 10% │
│             │             │             │                     │                  │ 2.55ms - 19.35ms │
├─────────────┼─────────────┼─────────────┼─────────────────────┼──────────────────┼──────────────────┤
│ tip-of-tree │ tip-of-tree │ 1026.74 KiB │ 182.51ms - 195.39ms │           slower │                  │
│             │             │             │                     │         1% - 11% │         -        │
│             │             │             │                     │ 2.55ms - 19.35ms │                  │
└─────────────┴─────────────┴─────────────┴─────────────────────┴──────────────────┴──────────────────┘

```